requirements.txt
requests
beautifulsoup4
tqdm
python-docx
rapidfuzz
